[language/e2e_tests] centralize account picking logic in one place
Going to add alternate means of picking accounts in the next diff.
